Remember the last chord event value
Save the "every Nth pixel" value used to paint chord events so
that we can skip that pixel when we draw other events on the next
paint cycle.  This allows chord events to stick around on the strip
longer.  It also gives more intensity (more LEDs lit) to higher chord
notes.